exampleapp: InMemory store is now moved in exampleapp
The InMemoryStore is now moved in the exampleapp. It's more reasonable
to be there, cause it's used in there and is not part of the core.
