+++ oauth2-example-app/src/main/java/com/clouway/oauth2/exampleapp/storage/InMemoryResourceOwnerRepository.java
@@ -1,4 +1,4 @@
-package com.example.auth.memory;
+package com.clouway.oauth2.exampleapp.storage;
 
 import com.clouway.oauth2.ResourceOwner;
 import com.clouway.oauth2.ResourceOwnerAuthentication;
