+++ oauth2-example-app/src/test/java/com/clouway/oauth2/exampleapp/storage/InMemoryClientRepositoryTest.java
@@ -1,7 +1,7 @@
-package com.example.auth.memory;
+package com.clouway.oauth2.exampleapp.storage;
 
 import com.clouway.oauth2.client.ClientRepository;
-import com.clouway.oauth2.client.ClientRepositoryContractTest;
+import com.clouway.oauth2.exampleapp.ClientRepositoryContractTest;
 
 /**
  * @author Ivan Stefanov <ivan.stefanov@clouway.com>
