+++ oauth2-example-app/src/test/java/com/clouway/oauth2/exampleapp/storage/InMemoryResourceOwnerRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.example.auth.memory;
+package com.clouway.oauth2.exampleapp.storage;
 
 import com.clouway.oauth2.ResourceOwner;
 import com.clouway.oauth2.Session;
